+++ tart-logging/src/commonMain/kotlin/io/yumemi/tart/logging/DefaultLogger.kt
@@ -3,7 +3,9 @@ package io.yumemi.tart.logging
 import co.touchlab.kermit.Severity
 import co.touchlab.kermit.Logger.Companion as Kermit
 
-class DefaultLogger : Logger {
+object DefaultLogger : Logger {
+    private var isDisabled = false
+
     override suspend fun log(severity: Logger.Severity, tag: String, throwable: Throwable?, message: String) {
         if (isDisabled) return
         Kermit.log(
@@ -14,6 +16,11 @@ class DefaultLogger : Logger {
         )
     }
 
+    @Suppress("unused")
+    fun disable() {
+        isDisabled = true
+    }
+
     private fun Logger.Severity.map() = when (this) {
         Logger.Severity.Verbose -> Severity.Verbose
         Logger.Severity.Debug -> Severity.Debug
@@ -21,14 +28,5 @@ class DefaultLogger : Logger {
         Logger.Severity.Warn -> Severity.Warn
         Logger.Severity.Error -> Severity.Error
         Logger.Severity.Assert -> Severity.Assert
-    }
-
-    companion object {
-        private var isDisabled = false
-
-        @Suppress("unused")
-        fun disable() {
-            isDisabled = true
-        }
     }
 }
