+++ tart-message/src/commonMain/kotlin/io/yumemi/tart/message/MessageMiddleware.kt
@@ -6,21 +6,8 @@ import io.yumemi.tart.core.Middleware
 import io.yumemi.tart.core.State
 import io.yumemi.tart.core.Store
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.launch
 import kotlin.coroutines.CoroutineContext
-
-interface Message
-
-private object MessageHub {
-    private val _messages = MutableSharedFlow<Message>()
-    val messages: Flow<Message> get() = _messages
-
-    suspend fun send(message: Message) {
-        _messages.emit(message)
-    }
-}
 
 @Suppress("unused")
 class MessageMiddleware<S : State, A : Action, E : Event>(
@@ -33,9 +20,4 @@ class MessageMiddleware<S : State, A : Action, E : Event>(
             }
         }
     }
-}
-
-@Suppress("unused")
-suspend fun send(message: Message) {
-    MessageHub.send(message = message)
 }
